Ensure all issues under a bottom-up item are excluded when bottom-up is excluded
Right now, we're only hiding bottom-up items if the item itself is marked as
bottom-up work. In other words, if the parent is bottom-up, but the child
isn't, then we use the skipping behavior that we use for themes/epics/user
stories, that is the item will show in the first ancestor that isn't
bottom-up. That isn't desirable.

Instead, when a part of the tree is marked as bottom-up, then we want to hide
that item and transitively all children, not matter whether a child isn't
marked as bottom-up or not.