Change filter behavior to show children of selected nodes
This now shows nodes if any of their children or parents are visible. If the
node itself isn't included in the filter, it's grayed out.

For example, this allows to quickly find a node in order to lookup their
children. Also, when setting the filter for the team it will also include
children if they aren't tagged with a team.