Mark completed and cut Azure DevOps items as closed
It seems on the Azure DevOps side there is no single "closed" state for these
kind of issues. So instead, we're saying an item should be considered closed
if it's either marked as completed or cut.

On the GitHub side, that's not the case as completed and cut are function of
the columns in the project, but in practice a human close this manually.